Preventing Field's mark overwriting
Now the Field can control what marks are specified. It usually can be only empty string or the specific characters of the tic-tac-toe game ("X" and "O").

When you try to use some other characters or set of characters longer than 1 character, or when you try to overwrite already specified mark, it results in an exception.